optional rust validation
